+++ src/client.js
@@ -30,6 +30,31 @@ function isUrl(url) {
   return SIMPLE_HTTP_URL_REGEX.test(url);
 }
 
+const isUnvalidId = (id) =>
+  !_.isUndefined(id) && !AGENT_ID_ALLOWED_REGEXP.test(id);
+const isUnvalidConfiguration = (configuration) =>
+  _.isUndefined(configuration) || !_.isObject(configuration);
+const areUnvalidOperations = (operations) =>
+  _.isUndefined(operations) || !_.isArray(operations);
+
+function checkBulkParameters(bulkArray) {
+  if (_.isUndefined(bulkArray)) {
+    throw new CraftAiBadRequestError(
+      'Bad Request, unable to use bulk functionalities without list provided.'
+    );
+  }
+  if (!_.isArray(bulkArray)) {
+    throw new CraftAiBadRequestError(
+      'Bad Request, bulk inputs should be provided within an array.'
+    );
+  }
+  if (!bulkArray.length) {
+    throw new CraftAiBadRequestError(
+      'Bad Request, the array containing bulk inputs is empty.'
+    );
+  }
+}
+
 export default function createClient(tokenOrCfg) {
   let cfg = _.defaults(
     {},
@@ -39,7 +64,9 @@ export default function createClient(tokenOrCfg) {
 
   // Initialization check
   if (!_.has(cfg, 'token') || !_.isString(cfg.token)) {
-    throw new CraftAiBadRequestError('Bad Request, unable to create a client with no or invalid token provided.');
+    throw new CraftAiBadRequestError(
+      'Bad Request, unable to create a client with no or invalid token provided.'
+    );
   }
   try {
     const { owner, platform, project } = jwtDecode(cfg.token);
@@ -53,11 +80,15 @@ export default function createClient(tokenOrCfg) {
     throw new CraftAiCredentialsError();
   }
   if (!_.has(cfg, 'url') || !isUrl(cfg.url)) {
-    throw new CraftAiBadRequestError('Bad Request, unable to create a client with no or invalid url provided.');
+    throw new CraftAiBadRequestError(
+      'Bad Request, unable to create a client with no or invalid url provided.'
+    );
   }
   if (!_.has(cfg, 'project') || !_.isString(cfg.project)) {
-    throw new CraftAiBadRequestError('Bad Request, unable to create a client with no or invalid project provided.');
-  }
+    throw new CraftAiBadRequestError(
+      'Bad Request, unable to create a client with no or invalid project provided.'
+    );
+  } 
   else {
     const splittedProject = cfg.project.split('/');
     if (splittedProject.length >= 2) {
@@ -66,13 +97,21 @@ export default function createClient(tokenOrCfg) {
     }
   }
   if (!_.has(cfg, 'owner') || !_.isString(cfg.owner)) {
-    throw new CraftAiBadRequestError('Bad Request, unable to create a client with no or invalid owner provided.');
+    throw new CraftAiBadRequestError(
+      'Bad Request, unable to create a client with no or invalid owner provided.'
+    );
   }
   if (cfg.proxy != null && !isUrl(cfg.proxy)) {
-    throw new CraftAiBadRequestError('Bad Request, unable to create a client with an invalid proxy url provided.');
+    throw new CraftAiBadRequestError(
+      'Bad Request, unable to create a client with an invalid proxy url provided.'
+    );
   }
 
-  debug(`Creating a client instance for project '${cfg.owner}/${cfg.project}' on '${cfg.url}'.`);
+  debug(
+    `Creating a client instance for project '${cfg.owner}/${cfg.project}' on '${
+      cfg.url
+    }'.`
+  );
 
   const request = createRequest(cfg);
 
@@ -80,12 +119,20 @@ export default function createClient(tokenOrCfg) {
   let instance = {
     cfg: cfg,
     createAgent: function(configuration, id = undefined) {
-      if (_.isUndefined(configuration) || !_.isObject(configuration)) {
-        return Promise.reject(new CraftAiBadRequestError('Bad Request, unable to create an agent with no or invalid configuration provided.'));
+      if (isUnvalidConfiguration(configuration)) {
+        return Promise.reject(
+          new CraftAiBadRequestError(
+            'Bad Request, unable to create an agent with no or invalid configuration provided.'
+          )
+        );
       }
 
-      if (!_.isUndefined(id) && !AGENT_ID_ALLOWED_REGEXP.test(id)) {
-        return Promise.reject(new CraftAiBadRequestError(`Bad Request, unable to create an agent with invalid agent id. It must only contain characters in "a-zA-Z0-9_-" and must be a string between 1 and ${AGENT_ID_MAX_LENGTH} characters.`));
+      if (isUnvalidId(id)) {
+        return Promise.reject(
+          new CraftAiBadRequestError(
+            `Bad Request, unable to create an agent with invalid agent id. It must only contain characters in 'a-zA-Z0-9_-' and must be a string between 1 and ${AGENT_ID_MAX_LENGTH} characters.`
+          )
+        );
       }
 
       return request({
@@ -101,9 +148,23 @@ export default function createClient(tokenOrCfg) {
           return body;
         });
     },
+    createAgentBulk: function(agentsList) {
+      checkBulkParameters(agentsList);
+
+      return request({
+        method: 'POST',
+        path: '/bulk/agents',
+        body: agentsList
+      })
+        .then(({ body }) => body);
+    },
     getAgent: function(agentId) {
       if (!AGENT_ID_ALLOWED_REGEXP.test(agentId)) {
-        return Promise.reject(new CraftAiBadRequestError('Bad Request, unable to get an agent with invalid agent id. It must only contain characters in "a-zA-Z0-9_-" and cannot be the empty string.'));
+        return Promise.reject(
+          new CraftAiBadRequestError(
+            'Bad Request, unable to get an agent with invalid agent id. It must only contain characters in \'a-zA-Z0-9_-\' and cannot be the empty string.'
+          )
+        );
       }
 
       return request({
@@ -121,7 +182,11 @@ export default function createClient(tokenOrCfg) {
     },
     deleteAgent: function(agentId) {
       if (_.isUndefined(agentId)) {
-        return Promise.reject(new CraftAiBadRequestError('Bad Request, unable to delete an agent with no agentId provided.'));
+        return Promise.reject(
+          new CraftAiBadRequestError(
+            'Bad Request, unable to delete an agent with no agentId provided.'
+          )
+        );
       }
 
       return request({
@@ -133,17 +198,35 @@ export default function createClient(tokenOrCfg) {
           return body;
         });
     },
+    deleteAgentBulk: function(agentsList) {
+      checkBulkParameters(agentsList);
+
+      return request({
+        method: 'DELETE',
+        path: '/bulk/agents',
+        body: agentsList
+      })
+        .then(({ body }) => body);
+    },
     destroyAgent: function(agentId) {
       deprecation('client.destroyAgent', 'client.deleteAgent');
       return this.deleteAgent(agentId);
     },
     getAgentContext: function(agentId, t = undefined) {
       if (_.isUndefined(agentId)) {
-        return Promise.reject(new CraftAiBadRequestError('Bad Request, unable to get the agent context with no agentId provided.'));
+        return Promise.reject(
+          new CraftAiBadRequestError(
+            'Bad Request, unable to get the agent context with no agentId provided.'
+          )
+        );
       }
       let posixTimestamp = Time(t).timestamp;
       if (_.isUndefined(posixTimestamp)) {
-        return Promise.reject(new CraftAiBadRequestError('Bad Request, unable to get the agent context with an invalid timestamp provided.'));
+        return Promise.reject(
+          new CraftAiBadRequestError(
+            'Bad Request, unable to get the agent context with an invalid timestamp provided.'
+          )
+        );
       }
 
       return request({
@@ -157,7 +240,11 @@ export default function createClient(tokenOrCfg) {
     },
     addAgentContextOperations: function(agentId, operations) {
       if (_.isUndefined(agentId)) {
-        return Promise.reject(new CraftAiBadRequestError('Bad Request, unable to add agent context operations with no agentId provided.'));
+        return Promise.reject(
+          new CraftAiBadRequestError(
+            'Bad Request, unable to add agent context operations with no agentId provided.'
+          )
+        );
       }
       if (!_.isArray(operations)) {
         // Only one given operation
@@ -166,7 +253,9 @@ export default function createClient(tokenOrCfg) {
       operations = _.compact(operations);
 
       if (!operations.length) {
-        const message = `No operation to add to the agent ${cfg.owner}/${cfg.project}/${agentId} context.`;
+        const message = `No operation to add to the agent ${cfg.owner}/${
+          cfg.project
+        }/${agentId} context.`;
 
         debug(message);
 
@@ -180,36 +269,123 @@ export default function createClient(tokenOrCfg) {
         }))
         .orderBy('timestamp')
         .chunk(cfg.operationsChunksSize)
-        .reduce((p, chunk) => p.then(
-          () => request({
-            method: 'POST',
-            path: `/agents/${agentId}/context`,
-            body: chunk
-          })
-        ),
-        Promise.resolve())
+        .reduce(
+          (p, chunk) =>
+            p.then(() =>
+              request({
+                method: 'POST',
+                path: `/agents/${agentId}/context`,
+                body: chunk
+              })
+            ),
+          Promise.resolve()
+        )
         .then(() => {
-          const message = `Successfully added ${operations.length} operation(s) to the agent ${cfg.owner}/${cfg.project}/${agentId} context.`;
+          const message = `Successfully added ${
+            operations.length
+          } operation(s) to the agent ${cfg.owner}/${
+            cfg.project
+          }/${agentId} context.`;
           debug(message);
           return { message };
         });
     },
-    getAgentContextOperations: function(agentId, start = undefined, end = undefined) {
+    addAgentContextOperationsBulk: function(agentsOperationsList) {
+      checkBulkParameters(agentsOperationsList);
+      agentsOperationsList.map(({ id, operations }) => {
+        if (areUnvalidOperations(operations)) {
+          throw new CraftAiBadRequestError(
+            `Bad Request, unable to handle operations for agent ${id}. Operations should be provided within an array.`
+          );
+        }
+      });
+
+      let chunkedData = [];
+      let currentChunk = [];
+      let currentChunkSize = 0;
+
+      for (let agent of agentsOperationsList) {
+        if (agent.operations && _.isArray(agent.operations)) {
+          if (
+            currentChunkSize + agent.operations.length >
+              cfg.operationsChunksSize &&
+            currentChunkSize.length
+          ) {
+            chunkedData.push(currentChunk);
+            currentChunkSize = 0;
+            currentChunk = [];
+          }
+
+          if (agent.operations.length > cfg.operationsChunksSize) {
+            chunkedData.push([agent]);
+            currentChunkSize = 0;
+          } 
+          else {
+            currentChunkSize += agent.operations.length;
+            currentChunk.push(agent);
+          }
+        }
+      }
+
+      if (currentChunk.length) {
+        chunkedData.push(currentChunk);
+      }
+
+      return Promise.all(
+        chunkedData.map((chunk) => {
+          if (chunk.length > 1) {
+            return request({
+              method: 'POST',
+              path: '/bulk/context',
+              body: chunk
+            })
+              .then(({ body }) => body);
+          } 
+          else {
+            return this.addAgentContextOperations(
+              chunk[0].id,
+              chunk[0].operations
+            )
+              .then(({ message }) => [
+                { id: chunk[0].id, status: 201, message }
+              ]);
+          }
+        })
+      )
+        .then(_.flattenDeep);
+    },
+    getAgentContextOperations: function(
+      agentId,
+      start = undefined,
+      end = undefined
+    ) {
       if (_.isUndefined(agentId)) {
-        return Promise.reject(new CraftAiBadRequestError('Bad Request, unable to get agent context operations with no agentId provided.'));
+        return Promise.reject(
+          new CraftAiBadRequestError(
+            'Bad Request, unable to get agent context operations with no agentId provided.'
+          )
+        );
       }
       let startTimestamp;
       if (start) {
         startTimestamp = Time(start).timestamp;
         if (_.isUndefined(startTimestamp)) {
-          return Promise.reject(new CraftAiBadRequestError('Bad Request, unable to get agent context operations with an invalid \'start\' timestamp provided.'));
+          return Promise.reject(
+            new CraftAiBadRequestError(
+              'Bad Request, unable to get agent context operations with an invalid \'start\' timestamp provided.'
+            )
+          );
         }
       }
       let endTimestamp;
       if (end) {
         endTimestamp = Time(end).timestamp;
         if (_.isUndefined(endTimestamp)) {
-          return Promise.reject(new CraftAiBadRequestError('Bad Request, unable to get agent context operations with an invalid \'end\' timestamp provided.'));
+          return Promise.reject(
+            new CraftAiBadRequestError(
+              'Bad Request, unable to get agent context operations with an invalid \'end\' timestamp provided.'
+            )
+          );
         }
       }
 
@@ -218,10 +394,13 @@ export default function createClient(tokenOrCfg) {
           return Promise.resolve(operations);
         }
         return request({ url: nextPageUrl }, this)
-          .then(({ body, nextPageUrl }) => requestFollowingPages({
-            operations: operations.concat(body),
-            nextPageUrl
-          }));
+          .then(
+            ({ body, nextPageUrl }) =>
+              requestFollowingPages({
+                operations: operations.concat(body),
+                nextPageUrl
+              })
+          );
       };
 
       return request({
@@ -232,27 +411,45 @@ export default function createClient(tokenOrCfg) {
           end: endTimestamp
         }
       })
-        .then(({ body, nextPageUrl }) => requestFollowingPages({
-          operations: body,
-          nextPageUrl
-        }));
+        .then(({ body, nextPageUrl }) =>
+          requestFollowingPages({
+            operations: body,
+            nextPageUrl
+          })
+        );
     },
-    getAgentStateHistory: function(agentId, start = undefined, end = undefined) {
+    getAgentStateHistory: function(
+      agentId,
+      start = undefined,
+      end = undefined
+    ) {
       if (_.isUndefined(agentId)) {
-        return Promise.reject(new CraftAiBadRequestError('Bad Request, unable to get agent state history with no agentId provided.'));
+        return Promise.reject(
+          new CraftAiBadRequestError(
+            'Bad Request, unable to get agent state history with no agentId provided.'
+          )
+        );
       }
       let startTimestamp;
       if (start) {
         startTimestamp = Time(start).timestamp;
         if (_.isUndefined(startTimestamp)) {
-          return Promise.reject(new CraftAiBadRequestError('Bad Request, unable to get agent state history with an invalid \'start\' timestamp provided.'));
+          return Promise.reject(
+            new CraftAiBadRequestError(
+              'Bad Request, unable to get agent state history with an invalid \'start\' timestamp provided.'
+            )
+          );
         }
       }
       let endTimestamp;
       if (end) {
         endTimestamp = Time(end).timestamp;
         if (_.isUndefined(endTimestamp)) {
-          return Promise.reject(new CraftAiBadRequestError('Bad Request, unable to get agent state history with an invalid \'end\' timestamp provided.'));
+          return Promise.reject(
+            new CraftAiBadRequestError(
+              'Bad Request, unable to get agent state history with an invalid \'end\' timestamp provided.'
+            )
+          );
         }
       }
 
@@ -261,10 +458,12 @@ export default function createClient(tokenOrCfg) {
           return Promise.resolve(stateHistory);
         }
         return request({ url: nextPageUrl })
-          .then(({ body, nextPageUrl }) => requestFollowingPages({
-            stateHistory: stateHistory.concat(body),
-            nextPageUrl
-          }));
+          .then(({ body, nextPageUrl }) =>
+            requestFollowingPages({
+              stateHistory: stateHistory.concat(body),
+              nextPageUrl
+            })
+          );
       };
 
       return request({
@@ -275,13 +474,18 @@ export default function createClient(tokenOrCfg) {
           end: endTimestamp
         }
       })
-        .then(({ body, nextPageUrl }) => requestFollowingPages({
-          stateHistory: body,
-          nextPageUrl
-        }));
+        .then(({ body, nextPageUrl }) =>
+          requestFollowingPages({
+            stateHistory: body,
+            nextPageUrl
+          })
+        );
     },
     getAgentInspectorUrl: function(agentId, t = undefined) {
-      deprecation('client.getAgentInspectorUrl', 'client.getSharedAgentInspectorUrl');
+      deprecation(
+        'client.getAgentInspectorUrl',
+        'client.getSharedAgentInspectorUrl'
+      );
       return this.getSharedAgentInspectorUrl(agentId, t);
     },
     getSharedAgentInspectorUrl: function(agentId, t = undefined) {
@@ -305,29 +509,42 @@ export default function createClient(tokenOrCfg) {
         path: `/agents/${agentId}/shared`
       })
         .then(() => {
-          debug(`Delete shared inspector link for agent "${agentId}".`);
+          debug(`Delete shared inspector link for agent '${agentId}'.`);
         });
     },
-    getAgentDecisionTree: function(agentId, t = undefined, version = DEFAULT_DECISION_TREE_VERSION) {
+    getAgentDecisionTree: function(
+      agentId,
+      t = undefined,
+      version = DEFAULT_DECISION_TREE_VERSION
+    ) {
       if (_.isUndefined(agentId)) {
-        return Promise.reject(new CraftAiBadRequestError('Bad Request, unable to retrieve an agent decision tree with no agentId provided.'));
+        return Promise.reject(
+          new CraftAiBadRequestError(
+            'Bad Request, unable to retrieve an agent decision tree with no agentId provided.'
+          )
+        );
       }
       let posixTimestamp = Time(t).timestamp;
       if (_.isUndefined(posixTimestamp)) {
-        return Promise.reject(new CraftAiBadRequestError('Bad Request, unable to retrieve an agent decision tree with an invalid timestamp provided.'));
+        return Promise.reject(
+          new CraftAiBadRequestError(
+            'Bad Request, unable to retrieve an agent decision tree with an invalid timestamp provided.'
+          )
+        );
       }
 
-      const agentDecisionTreeRequest = () => request({
-        method: 'GET',
-        path: `/agents/${agentId}/decision/tree`,
-        query: {
-          t: posixTimestamp
-        },
-        headers: {
-          'x-craft-ai-tree-version': version
-        }
-      })
-        .then(({ body }) => body);
+      const agentDecisionTreeRequest = () =>
+        request({
+          method: 'GET',
+          path: `/agents/${agentId}/decision/tree`,
+          query: {
+            t: posixTimestamp
+          },
+          headers: {
+            'x-craft-ai-tree-version': version
+          }
+        })
+          .then(({ body }) => body);
 
       if (!cfg.decisionTreeRetrievalTimeout) {
         // Don't retry
@@ -340,12 +557,20 @@ export default function createClient(tokenOrCfg) {
             .catch((error) => {
               const requestDuration = Date.now() - start;
               const expectedRetryDuration = requestDuration + 2000; // Let's add some margin
-              const timeoutBeforeRetrying = cfg.decisionTreeRetrievalTimeout - requestDuration - expectedRetryDuration;
-              if (error instanceof CraftAiLongRequestTimeOutError && timeoutBeforeRetrying > 0) {
+              const timeoutBeforeRetrying =
+                cfg.decisionTreeRetrievalTimeout -
+                requestDuration -
+                expectedRetryDuration;
+              if (
+                error instanceof CraftAiLongRequestTimeOutError &&
+                timeoutBeforeRetrying > 0
+              ) {
                 // First timeout, let's retry once near the end of the set timeout
                 return resolveAfterTimeout(timeoutBeforeRetrying)
-                  .then(() => agentDecisionTreeRequest());
-              }
+                  .then(() =>
+                    agentDecisionTreeRequest()
+                  );
+              } 
               else {
                 return Promise.reject(error);
               }
@@ -357,16 +582,38 @@ export default function createClient(tokenOrCfg) {
         ]);
       }
     },
+    getAgentDecisionTreeBulk: function(agentsList) {
+      checkBulkParameters(agentsList);
+
+      return request({
+        method: 'POST',
+        path: '/bulk/decision_tree',
+        body: agentsList
+      })
+        .then(({ body }) => body);
+    },
     computeAgentDecision: function(agentId, t, ...contexts) {
       if (_.isUndefined(agentId)) {
-        return Promise.reject(new CraftAiBadRequestError('Bad Request, unable to compute an agent decision with no agentId provided.'));
+        return Promise.reject(
+          new CraftAiBadRequestError(
+            'Bad Request, unable to compute an agent decision with no agentId provided.'
+          )
+        );
       }
       let posixTimestamp = Time(t).timestamp;
       if (_.isUndefined(posixTimestamp)) {
-        return Promise.reject(new CraftAiBadRequestError('Bad Request, unable to compute an agent decision with no or invalid timestamp provided.'));
+        return Promise.reject(
+          new CraftAiBadRequestError(
+            'Bad Request, unable to compute an agent decision with no or invalid timestamp provided.'
+          )
+        );
       }
       if (_.isUndefined(contexts) || _.size(contexts) === 0) {
-        return Promise.reject(new CraftAiBadRequestError('Bad Request, unable to compute an agent decision with no context provided.'));
+        return Promise.reject(
+          new CraftAiBadRequestError(
+            'Bad Request, unable to compute an agent decision with no context provided.'
+          )
+        );
       }
 
       return request({